repo/requirements.txt
beautifulsoup4==4.12.3
pandas==1.5.3
requests==2.31.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.27
alembic==1.13.1
GeoAlchemy2==0.14.3
python-dotenv==1.0.1 